+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 # ansibleHelper Changelog
 ## [Unreleased]
+### Added
+- ansible json schema's for validation
+- Added json schema's for Ansible
 
 ### Fixed
 - changelog updates to published plugin so it can be seen within idea marketplace